+++ src/controller/AppConst.js
@@ -1,0 +1,3 @@
+// module.exports.AppConst = {
+//   DEPLOY_URL: ''
+// };