+++ src/controller/ExpressAppController.js
@@ -17,6 +17,7 @@ class ExpressAppController {
     /**
      *
      * @param functions {{
+     *     path: string,
      *     onRequest: [Function | [Function, Function, ...] | Router]
      * }}
      * @param port
@@ -28,26 +29,40 @@ class ExpressAppController {
 
     start() {
         if (typeof this._functions === 'object') {
-            _app.use('/names', (request, response) => {
-                let functionsNames = [];
-                Object.keys(this._functions).forEach(functionName => {
-                    if (this._functions[functionName] && typeof this._functions[functionName] === "object"
-                        && this._functions[functionName].onRequest) {
-                        functionsNames.push(functionName)
-                    }
-                });
-                response.json({names: functionsNames});
-            });
+            /* handle return of all functions registered */
+            // _app.use('/names', (request, response) => {
+            //
+            //
+            //     _app._router.stack.forEach(layer => {
+            //         console.log(layer);
+            //         console.log(layer.name);
+            //         console.log(layer.path);
+            //     });
+            //
+            //     let functionsNames = [];
+            //     Object.keys(this._functions).forEach(functionName => {
+            //         if (this._functions[functionName] && typeof this._functions[functionName] === "object"
+            //             && this._functions[functionName].onRequest) {
+            //             functionsNames.push(functionName)
+            //         }
+            //     });
+            //     response.json({names: functionsNames});
+            // });
+
             Object.keys(this._functions).forEach(functionName => {
                 if (this._functions[functionName] && typeof this._functions[functionName] === "object"
                     && this._functions[functionName].onRequest) {
-                    _app.use(`/functions/${functionName}`, this._functions[functionName].onRequest);
+                    if (this._functions[functionName].path) {
+                        _app.use(this._functions[functionName].path, this._functions[functionName].onRequest);
+                    } else {
+                        _app.use(`/functions/${functionName}`, this._functions[functionName].onRequest);
+                    }
                 }
             });
             const faasServer = http.createServer(_app);
             faasServer.listen(this._port.toString());
             faasServer.on('listening', () => {
-                console.log('FaaS Engine Listening on ' + this._port);
+                console.log('BFast::Cloud Functions engine listening on ' + this._port);
             });
         } else {
             throw {message: 'It\'s not object, hence functions no served'};