+++ src/controller/FunctionController.js
@@ -1,17 +1,15 @@
 const glob = require('glob');
-const https = require('https');
+const axios = require('axios');
 const ExpressApp = require('./ExpressAppController');
+const DatabaseController = require('./DatabaseController');
+const ResourceFactory = require('../factory/ResourceFactory');
+
+const _storage = new DatabaseController();
+const _resourceFactory = new ResourceFactory();
 
 class FunctionController {
 
-    /**
-     *
-     * @param resourceFactory {ResourceFactory}
-     * @param authController {AuthController}
-     */
-    constructor({resourceFactory, authController}) {
-        this._resourceFactory = resourceFactory;
-        this._authController = authController;
+    constructor() {
         this._getFunctions = (projectDir) => {
             let bfastJSON = require(`${projectDir}/bfast.json`);
             const bfastStringfied = JSON.stringify(bfastJSON);
@@ -21,10 +19,10 @@ class FunctionController {
                 ignore: bfastJSON.ignore
             });
             let functions = {
-                mambo: {
-                    onRequest: function (req, response) {
-                        response.json({message: 'Powa!'});
-                    }
+                _init: {
+                    // onRequest: function (req, response) {
+                    //     response.json({message: 'Powa!'});
+                    // }
                 }
             };
             files.forEach(file => {
@@ -46,24 +44,25 @@ class FunctionController {
          * @returns {Promise<{projectId}>}. Promise rejected when <code>bfast.json</code> is no found.
          * @private
          */
-        this._getProjectCredential = (projectDir) => {
+        this._checkIsBFastProjectFolder = (projectDir) => {
             return new Promise((resolve, reject) => {
                 try {
                     const projectCredential = require(`${projectDir}/bfast.json`);
-                    // console.log(`current project is : ${projectCredential.projectId}`);
-                    if (projectCredential && projectCredential.projectId && projectCredential.projectId !== '' &&
-                        projectCredential.projectId !== undefined && projectCredential.projectId !== null) {
+                    if (projectCredential && projectCredential.ignore) {
                         resolve(projectCredential);
                     } else {
                         reject('projectId can not be determined, ' +
                             'check if your current directory is bfast project and bfast.json file exist');
                     }
                 } catch (e) {
-                    // console.log(e);
                     reject('can not serve project, error : not in bfast project folder');
                 }
             });
         }
+
+        // this._checkProjectIsClean = () => {
+        //
+        // }
     }
 
     /**
@@ -71,8 +70,7 @@ class FunctionController {
      * @param projectDir {string} a workspace
      */
     initiateFunctionsFolder(projectDir) {
-        if (!this._authController.isAuthenticated()) throw Error('login first');
-        this._resourceFactory.createProjectFolder(projectDir)
+        return _resourceFactory.createProjectFolder(projectDir)
     }
 
     /**
@@ -80,18 +78,16 @@ class FunctionController {
      * @param projectDir
      * @param port
      */
-    // need to auto reload when file changes
     serve(projectDir, port) {
         try {
             const bfastFile = require(`${projectDir}/bfast.json`);
-            if (bfastFile && bfastFile.projectId !== undefined || bfastFile.projectId !== null) {
+            if (bfastFile && bfastFile.ignore) {
                 const functions = this._getFunctions(projectDir);
                 this._serveFunctions(functions, port);
             } else {
                 console.log('not in project folder or project file is invalid');
             }
         } catch (e) {
-            //  console.log(e);
             console.log('can not serve project, error : not in bfast project folder');
         }
     }
@@ -100,92 +96,113 @@ class FunctionController {
      * Deploy project to bfast cloud
      * @param projectDir
      * @param force
+     * @param options {{token: string, projectId: string}}
      */
-    // need to check if user is log in bfast project
-    deploy(projectDir, force = false) {
-        this._getProjectCredential(projectDir).then(bfastProject => {
-            https.request(
-                `https://cloud.bfast.fahamutech.com/deploy/functions/${bfastProject.projectId}?force=${force}`,
-                res => {
-                    res.setEncoding('utf8');
-                    // res.on('data', chunk => {
-                    //     console.log(chunk);
-                    // });
-                    res.on('error', err => {
-                        console.error(err);
-                    });
-                    res.on('end', (_) => {
-                        console.log('functions deployed');
-                    });
-                }).end();
-        }).catch(reason => {
-            console.log(reason);
-        });
-    }
-
-    addEnv(projectDir, envs, force = false) {
-        if (envs && Array.isArray(envs)) {
-            this._getProjectCredential(projectDir).then(bfastProject => {
-                console.log('start update faas environments');
-                const httpsReq = https.request(
-                    `https://cloud.bfast.fahamutech.com/functions/${bfastProject.projectId}/env?force=${force}`,
-                    {
-                        method: 'POST',
-                        headers: {
-                            "content-type": "application/json"
-                        }
-                    }, res => {
-                        res.setEncoding('utf8');
-                        res.on('data', chunk => {
-                            console.log(chunk);
-                        });
-                        res.on('error', err => {
-                            console.error(err);
-                        });
-                        res.on('end', (_) => {
-                            // console.log('env(s) added');
-                        });
-                    });
-                httpsReq.write(JSON.stringify({envs: envs}));
-                httpsReq.end();
-            }).catch(reason => {
-                console.log(reason);
-            });
-        } else {
-            console.log('No env(s) to update');
+    async deploy(projectDir, force = false, options) {
+        try {
+            await this._checkIsBFastProjectFolder(projectDir);
+            let projectId;
+            let token;
+            if (options && options.token && options.projectId) {
+                token = options.token;
+                projectId = options.projectId;
+            } else {
+                const user = await _storage.getUser();
+                const project = await _storage.getCurrentProject();
+                projectId = project.projectId;
+                token = user.token;
+            }
+            console.log(`\nCurrent linked bfast project ( projectId: ${projectId})`);
+            const response = await axios.post(
+                `https://api.bfast.fahamutech.com/functions/${projectId}?force=${force}`,
+                {},
+                {
+                    headers: {
+                        'authorization': `Bearer ${token}`,
+                    }
+                }
+            );
+            return response.data;
+        } catch (reason) {
+            if (reason && reason.response) {
+                throw reason.response.data;
+            } else if (reason.message) {
+                throw reason.message;
+            } else if (typeof reason === 'object') {
+                throw reason;
+            } else {
+                throw reason.toString();
+            }
         }
     }
 
-    removeEnv(projectDir, envs, force = false) {
+    async addEnv(projectDir, envs, force = false) {
         if (envs && Array.isArray(envs)) {
-            this._getProjectCredential(projectDir).then(bfastProject => {
-                console.log('start remove faas environments');
-                const httpsReq = https.request(
-                    `https://cloud.bfast.fahamutech.com/functions/${bfastProject.projectId}/env/delete?force=${force}`,
+            try {
+                await this._checkIsBFastProjectFolder(projectDir);
+                const user = await _storage.getUser();
+                const project = await _storage.getCurrentProject();
+                const projectId = project.projectId;
+                const token = user.token;
+                console.log(`\nCurrent linked bfast project ( projectId: ${projectId})`);
+                console.log('start add cloud functions environment(s)');
+                const response = await axios.post(
+                    `https://api.bfast.fahamutech.com/functions/${projectId}/env?force=${force}`,
                     {
-                        method: 'POST',
+                        envs: envs
+                    },
+                    {
                         headers: {
-                            "content-type": "application/json"
+                            'content-type': 'application/json',
+                            'authorization': `Bearer ${token}`
                         }
-                    }, res => {
-                        res.setEncoding('utf8');
-                        res.on('data', chunk => {
-                            console.log(chunk);
-                        });
-                        res.on('error', err => {
-                            console.error(err);
-                        });
-                        res.on('end', (_) => {
-                            // console.log('env(s) removed');
-                        });
-                    });
-                httpsReq.write(JSON.stringify({envs: envs}));
-                httpsReq.end();
-            }).catch(reason => {
-                console.log(reason);
-            });
+                    }
+                );
+                return response.data;
+            } catch (reason) {
+                if (reason && reason.response) {
+                    throw reason.response.data;
+                } else {
+                    throw reason.toString();
+                }
+            }
         } else {
-            console.log('No env(s) to update');
+            throw 'Please specify env(s) to add';
+        }
+    }
+
+    async removeEnv(projectDir, envs, force = false) {
+        try {
+            if (envs && Array.isArray(envs)) {
+                await this._checkIsBFastProjectFolder(projectDir);
+                const user = await _storage.getUser();
+                const project = await _storage.getCurrentProject();
+                const projectId = project.projectId;
+                const token = user.token;
+                console.log(`\nCurrent linked bfast project ( projectId: ${projectId})`);
+                console.log('start removing cloud functions environment(s)');
+                const response = await axios.delete(
+                    `https://api.bfast.fahamutech.com/functions/${projectId}/env?force=${force}`,
+                    {
+                        headers: {
+                            'content-type': 'application/json',
+                            'authorization': `Bearer ${token}`
+                        },
+                        data: {
+                            envs: envs
+                        }
+                    }
+                );
+                return response.data;
+            } else {
+                throw 'Please specify env(s) to remove';
+            }
+        } catch (reason) {
+            if (reason && reason.response) {
+                throw reason.response.data;
+            } else {
+                throw reason.toString();
+            }
         }
     }
 }