+++ src/controller/index.js
@@ -1,7 +1,0 @@
-const AuthController = require('./AuthController');
-const FunctionController = require('./FunctionController');
-
-module.exports = {
-    auth: AuthController,
-    functions: FunctionController
-};